Integrating interfaces into algorithms
